breakpseudo: Add demo of breaking pseudo with symlink creation
We sometimes see https://bugzilla.yoctoproject.org/show_bug.cgi?id=14957 on
the autobuilder. This recipe replicates that failure. Usually it occurs in
pciutils in creation of libpci.so.

The issue is the makefile can result in multiple calls to ln -sf in
parallel which race and confuse pseudo.

We have two issues:

a) it would be nice to fix the picutils makefiles so they don't race and call
this multiple times

b) pseudo should fail like this

I'm sharing this reproducer since it moves the bug from "I can't understand
or reproduce this" to a "there is a reproducer".

I commonly get complaints that bugs aren't reproducible, this one now is. I'm
hoping that means I will get some help on it.

To reprocude, run "bitbake breakpseudo" and watch the inoode mismatch errors.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>